refactor into authentication_provider package and refactor to use authentication_provider_factory.py
