+++ backend/ops_api/ops/auth/decorators.py
@@ -24,7 +24,8 @@ def is_user_active(f):
 
         from ops_api.ops.auth.authentication_gateway import AuthenticationGateway
 
-        auth_gateway = AuthenticationGateway(current_app.config.get("JWT_PRIVATE_KEY"))
+        # auth_gateway = AuthenticationGateway(current_app.config.get("JWT_PRIVATE_KEY"))
+        auth_gateway = AuthenticationGateway(current_app.config)
 
         provider = request.json.get("provider")
         user_info = auth_gateway.get_user_info(provider, token.get("access_token"))
