+++ backend/ops_api/ops/auth/service.py
@@ -19,7 +19,8 @@ def login(code: str, provider: str) -> dict[str, Any]:
     current_app.logger.debug(f"login - provider: {provider}")
 
     with current_app.app_context():
-        auth_gateway = AuthenticationGateway(current_app.config.get("JWT_PRIVATE_KEY"))
+        # auth_gateway = AuthenticationGateway(current_app.config.get("JWT_PRIVATE_KEY"))
+        auth_gateway = AuthenticationGateway(current_app.config)
 
     with OpsEventHandler(OpsEventType.LOGIN_ATTEMPT) as la:
         token = auth_gateway.authenticate(provider, code)
